+++ era-compiler-solidity/src/missing_libraries.rs
@@ -3,21 +3,21 @@
 //!
 
 use std::collections::BTreeMap;
-use std::collections::HashSet;
+use std::collections::BTreeSet;
 
 ///
 /// The missing libraries.
 ///
 pub struct MissingLibraries {
     /// The libraries.
-    pub contract_libraries: BTreeMap<String, HashSet<String>>,
+    pub contract_libraries: BTreeMap<String, BTreeSet<String>>,
 }
 
 impl MissingLibraries {
     ///
     /// A shortcut constructor.
     ///
-    pub fn new(contract_libraries: BTreeMap<String, HashSet<String>>) -> Self {
+    pub fn new(contract_libraries: BTreeMap<String, BTreeSet<String>>) -> Self {
         Self { contract_libraries }
     }