+++ era-compiler-solidity/src/process/mod.rs
@@ -39,6 +39,7 @@ pub fn run(target: era_compiler_common::Target) -> anyhow::Result<()> {
                 .compile_to_eravm(
                     input.solc_version,
                     input.identifier_paths,
+                    input.missing_libraries,
                     input.factory_dependencies,
                     input.enable_eravm_extensions,
                     input.metadata_hash_type,