+++ era-yul/src/yul/parser/statement/assignment.rs
@@ -2,7 +2,7 @@
 //! The assignment expression statement.
 //!
 
-use std::collections::HashSet;
+use std::collections::BTreeSet;
 
 use crate::yul::error::Error;
 use crate::yul::lexer::token::lexeme::symbol::Symbol;
@@ -106,7 +106,7 @@ impl Assignment {
     ///
     /// Get the list of missing deployable libraries.
     ///
-    pub fn get_missing_libraries(&self) -> HashSet<String> {
+    pub fn get_missing_libraries(&self) -> BTreeSet<String> {
         self.initializer.get_missing_libraries()
     }
 }