+++ era-yul/src/yul/parser/statement/if_conditional.rs
@@ -2,7 +2,7 @@
 //! The if-conditional statement.
 //!
 
-use std::collections::HashSet;
+use std::collections::BTreeSet;
 
 use crate::yul::error::Error;
 use crate::yul::lexer::token::location::Location;
@@ -54,7 +54,7 @@ where
     ///
     /// Get the list of missing deployable libraries.
     ///
-    pub fn get_missing_libraries(&self) -> HashSet<String> {
+    pub fn get_missing_libraries(&self) -> BTreeSet<String> {
         let mut libraries = self.condition.get_missing_libraries();
         libraries.extend(self.block.get_missing_libraries());
         libraries